undo git revert
