+++ pkg/scm/git/git_test.go
@@ -2,9 +2,6 @@ package git
 
 import (
 	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"reflect"
 	"testing"
 
@@ -12,104 +9,23 @@ import (
 	"github.com/openshift/source-to-image/pkg/test"
 )
 
-func createLocalGitDirectory(t *testing.T) string {
-	dir, err := ioutil.TempDir(os.TempDir(), "s2i-test")
-	if err != nil {
-		t.Error(err)
-	}
-	os.Mkdir(filepath.Join(dir, ".git"), 0600)
-	return dir
-}
-
-func TestIsLocalGitRepository(t *testing.T) {
-	d := createLocalGitDirectory(t)
-	defer os.RemoveAll(d)
-	if isLocalGitRepository(d) == false {
-		t.Errorf("The %q directory is git repository", d)
-	}
-	os.RemoveAll(filepath.Join(d, ".git"))
-	if isLocalGitRepository(d) == true {
-		t.Errorf("The %q directory is not git repository", d)
-	}
-}
-
 func TestValidCloneSpec(t *testing.T) {
-	gitLocalDir := createLocalGitDirectory(t)
-	defer os.RemoveAll(gitLocalDir)
-
-	valid := []string{"git@github.com:user/repo.git",
+	scenarios := []string{"git@github.com:user/repo.git",
 		"git://github.com/user/repo.git",
 		"git://github.com/user/repo",
 		"http://github.com/user/repo.git",
 		"http://github.com/user/repo",
 		"https://github.com/user/repo.git",
 		"https://github.com/user/repo",
-		"file://" + gitLocalDir,
-		gitLocalDir,
-		"git@192.168.122.1:repositories/authooks",
-		"mbalazs@build.ulx.hu:/var/git/eap-ulx.git",
-		"ssh://git@[2001:db8::1]/repository.git",
-		"ssh://git@mydomain.com:8080/foo/bar",
-		"git@[2001:db8::1]:repository.git",
-		"git@[2001:db8::1]:/repository.git",
-		"g_m@foo.bar:foo/bar",
-		"g-m@foo.bar:foo/bar",
-		"g.m@foo.bar:foo/bar",
-		"github.com:openshift/origin.git",
-		"http://github.com/user/repo#1234",
-	}
-	invalid := []string{"g&m@foo.bar:foo.bar",
-		"~git@test.server/repo.git",
-		"some/lo:cal/path", // a local path that does not exist, but "some/lo" is not a valid host name
-		"http://github.com/user/repo#%%%%",
-	}
-
-	gh := New()
-
-	for _, scenario := range valid {
-		result := gh.ValidCloneSpec(scenario)
-		if result == false {
-			t.Error(scenario)
-		}
-	}
-	for _, scenario := range invalid {
-		result := gh.ValidCloneSpec(scenario)
-		if result {
-			t.Error(scenario)
-		}
-	}
-}
-
-func TestValidCloneSpecRemoteOnly(t *testing.T) {
-	valid := []string{"git://github.com/user/repo.git",
-		"git://github.com/user/repo",
-		"http://github.com/user/repo.git",
-		"http://github.com/user/repo",
-		"https://github.com/user/repo.git",
-		"https://github.com/user/repo",
-		"ssh://git@[2001:db8::1]/repository.git",
-		"ssh://git@mydomain.com:8080/foo/bar",
-		"git@github.com:user/repo.git",
-		"git@192.168.122.1:repositories/authooks",
-		"mbalazs@build.ulx.hu:/var/git/eap-ulx.git",
-		"git@[2001:db8::1]:repository.git",
-		"git@[2001:db8::1]:/repository.git",
-	}
-	invalid := []string{"file:///home/user/code/repo.git",
+		"file:///home/user/code/repo.git",
 		"/home/user/code/repo.git",
 	}
 
 	gh := New()
 
-	for _, scenario := range valid {
-		result := gh.ValidCloneSpecRemoteOnly(scenario)
+	for _, scenario := range scenarios {
+		result := gh.ValidCloneSpec(scenario)
 		if result == false {
-			t.Error(scenario)
-		}
-	}
-	for _, scenario := range invalid {
-		result := gh.ValidCloneSpecRemoteOnly(scenario)
-		if result {
 			t.Error(scenario)
 		}
 	}
